# docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "This is a sample API documentation for the Totes backend using Gin and Swagger.",
        "title": "Totes Backend API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "email": "support@totesbackend.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost",
    "basePath": "/",
    "paths": {
...
        "/external-sales": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Fetches a list of all external sales, including related items and customers.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "external-sales"
                ],
                "summary": "Retrieve all external sales",
                "responses": {
                    "200": {
                        "description": "Successfully retrieved all external sales",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/dtos.GetExternalSaleDTO"
                            }
                        }
                    },
                    "403": {
                        "description": "Access denied",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Error retrieving external sales",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Creates a new external sale, including the sale details and customer information.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "external-sales"
                ],
                "summary": "Create a new external sale",
                "parameters": [
                    {
                        "description": "External Sale data",
                        "name": "external-sale",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dtos.CreateExternalSaleDTO"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Successfully created external sale",
                        "schema": {
                            "$ref": "#/definitions/dtos.GetExternalSaleDTO"
                        }
                    },
                    "400": {
                        "description": "Invalid JSON format",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Error creating external sale",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/external-sales/{id}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Fetches the details of a specific external sale by its ID.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "external-sales"
                ],
                "summary": "Retrieve external sale by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "External Sale ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully retrieved external sale",
                        "schema": {
                            "$ref": "#/definitions/dtos.GetExternalSaleDTO"
                        }
                    },
                    "403": {
                        "description": "Access denied",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "External sale not found",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Error retrieving external sale",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/historical-item-prices/{id}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Retrieves the historical prices for an item based on the item ID provided.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "historical-item-prices"
                ],
                "summary": "Get historical price for an item",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Item ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully retrieved historical prices",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.HistoricalItemPrice"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid Item ID",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "No historical prices found",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Failed to retrieve historical prices",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/identifier-types": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Retrieves a list of all available identifier types.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "identifier-types"
                ],
                "summary": "Get all identifier types",
                "responses": {
                    "200": {
                        "description": "Successfully retrieved identifier types",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.IdentifierType"
                            }
                        }
                    },
                    "403": {
                        "description": "Access denied",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Error retrieving identifier types",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/identifier-types/{id}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Retrieves an identifier type by its ID.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "identifier-types"
                ],
                "summary": "Get identifier type by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Identifier Type ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully retrieved identifier type",
                        "schema": {
                            "$ref": "#/definitions/models.IdentifierType"
                        }
                    },
                    "403": {
                        "description": "Access denied",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Identifier Type not found",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "dtos.CreateExternalSaleDTO": {
            "type": "object",
            "required": [
                "customerId",
                "customerName",
                "email",
                "identifierTypeId",
                "item_id",
                "lastName",
                "reporter_id",
                "reporter_name"
            ],
            "properties": {
                "address": {
                    "type": "string"
                },
                "customerId": {
                    "type": "string"
                },
                "customerName": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "identifierTypeId": {
                    "type": "integer"
                },
                "isBusiness": {
                    "type": "boolean"
                },
                "item_id": {
                    "type": "integer"
                },
                "lastName": {
                    "type": "string"
                },
                "phoneNumbers": {
                    "type": "string"
                },
                "reporter_id": {
                    "type": "string"
                },
                "reporter_name": {
                    "type": "string"
                },
                "stock": {
                    "type": "integer"
                }
            }
        },
...
        "dtos.GetExternalSaleDTO": {
            "type": "object",
            "properties": {
                "customer_email": {
                    "type": "string"
                },
                "customer_id": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "item_id": {
                    "type": "integer"
                },
                "item_name": {
                    "type": "string"
                },
                "reporter_id": {
                    "type": "string"
                },
                "reporter_name": {
                    "type": "string"
                },
                "stock": {
                    "type": "integer"
                }
            }
        },
...
        "models.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        },
        "models.HistoricalItemPrice": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "item_id": {
                    "type": "integer"
                },
                "modified_at": {
                    "type": "string"
                },
                "price": {
                    "type": "number"
                }
            }
        },
        "models.IdentifierType": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                }
            }
        },
...
    }
}